improve agent health deadlock behaviour
